Michalle

+++ src/pages/PetDetailsPage/PetDetailsPage.jsx
@@ -5,9 +5,10 @@ import { useEffect, useState } from "react";
 //import { useCommentsContext } from "../../hooks/useCommentsContext";
 import * as commentsAPI from "../../utilities/comments-api";
 import * as postsAPI from "../../utilities/posts-api";
+//import user from "../../../models/user";
 //import post from "../../../models/post";
 
-export default function PetDetails() {
+export default function PetDetails({ user }) {
   /*========================================
         Post Part
 ========================================*/
@@ -26,9 +27,30 @@ export default function PetDetails() {
   }, []);
 
   async function handleDeletePost () {
-    const del = await postsAPI.deletePost(postId)
+    if(thePost.user === user._id) {
+      const del = await postsAPI.deletePost(postId)
     console.log(del);
     window.location.href = `/AllPosts`;
+    }
+    else {
+      alert("cannot delete because you are not the user")
+    }
+    //       const del = await postsAPI.deletePost(postId)
+    // console.log(del);
+    // window.location.href = `/AllPosts`;
+  }
+
+  function handleEditPost () {
+    if(thePost.user !== user._id) {
+      alert("cannot edit because you are not the user")
+      window.location.href = `/${postId}`;
+    }
+    else {
+      window.location.href = `/${postId}/EditPost`
+    }
+    //       const del = await postsAPI.deletePost(postId)
+    // console.log(del);
+    // window.location.href = `/AllPosts`;
   }
 
   /*========================================
@@ -43,18 +65,10 @@ export default function PetDetails() {
     async function fetchComments() {
       const com = await commentsAPI.getAll(postId);
       setComments(com);
-      // dispatch({ type: "SET_COMMENTS", payload: com });
     }
     fetchComments();
   }, []);
 
-  // function addComment(comment) {
-  //   setComments({ ...comments, comment });
-  //   console.log(comments); //got array of comment objects
-  // }
-
-  let editURL = `/${postId}/EditPost`;
-  let deleteURL = `/${postId}`;
 
   return (
     <>
@@ -63,6 +77,7 @@ export default function PetDetails() {
         {thePost && (
           <>
             <p>Author: {thePost.userName}</p>
+            {/* <p>User: {thePost.user.name} </p> */}
             <p>Contact Info: {thePost.contactInfo}</p>
             <p>Post Type: {thePost.postType}</p>
             {/* Need to update time later --- */}
@@ -72,7 +87,7 @@ export default function PetDetails() {
             <p>Animal Name:{thePost.name}</p>
             <p>Animal Type:{thePost.animalType} </p>
             <p>Images: </p>
-            <img src={thePost.imgURL} alt={thePost.name} />
+            <img src={thePost.imgURL} alt={thePost.name} width="200px"/>
             <p>Animal Age: {thePost.age}</p>
             <p>Last Seen Location: </p>
             <p>reserved place for map api</p>
@@ -88,9 +103,7 @@ export default function PetDetails() {
         <br />
         <br />
         {/* </div>let petURL = `/${post._id}`; */}
-        <a href={editURL}>
-          <button>Edit</button>
-        </a>
+          <button onClick={handleEditPost}>Edit</button>
           <button onClick={handleDeletePost}>Delete</button>
       </div>
       {/* Is there any comments? comments.length -not works every time?! */}
